# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "All HTTP + MQ endpoints exposed by the goop2 viewer.\\n\\nAll peer-to-peer signaling travels through MQ (POST /api/mq/send → P2P → /api/mq/events SSE).\\nSee x-mq-topics in the raw spec (/api/openapi.json) for the full topic contract.",
        "title": "goop2 Viewer API",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/docs/browse": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "docs"
                ],
                "summary": "Aggregate file lists from all group members (parallel fetch)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "group_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/docs/delete": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "docs"
                ],
                "summary": "Delete a shared file (local access only)",
                "parameters": [
                    {
                        "description": "Delete request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.docsDeleteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.statusOK"
                        }
                    }
                }
            }
        },
        "/api/docs/download": {
            "get": {
                "tags": [
                    "docs"
                ],
                "summary": "Download a file (local store or proxied from remote peer)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "group_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Filename",
                        "name": "file",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Peer ID (empty = self)",
                        "name": "peer_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pass '1' for Content-Disposition: inline",
                        "name": "inline",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File content",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/docs/my": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "docs"
                ],
                "summary": "List my shared files for a group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "group_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/routes.docFileInfo"
                            }
                        }
                    }
                }
            }
        },
        "/api/docs/upload": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "docs"
                ],
                "summary": "Upload a file to share with the group (multipart)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "group_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "File to upload (max 50 MB)",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.statusOK"
                        }
                    }
                }
            }
        },
        "/api/docs/upload-local": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "docs"
                ],
                "summary": "Upload a file from a local filesystem path to share with the group",
                "parameters": [
                    {
                        "description": "Upload request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.docsUploadLocalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.statusOK"
                        }
                    },
                    "400": {
                        "description": "Missing group_id or path / Cannot read file",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "routes.docFileInfo": {
            "type": "object",
            "properties": {
                "mod_time": {
                    "type": "string",
                    "example": "2026-03-08T12:00:00Z"
                },
                "name": {
                    "type": "string",
                    "example": "notes.pdf"
                },
                "size": {
                    "type": "integer",
                    "example": 1048576
                }
            }
        },
...
        "routes.docsDeleteRequest": {
            "type": "object",
            "properties": {
                "filename": {
                    "type": "string",
                    "example": "notes.pdf"
                },
                "group_id": {
                    "type": "string",
                    "example": "a1b2c3d4e5f6a1b2"
                }
            }
        },
        "routes.docsUploadLocalRequest": {
            "type": "object",
            "properties": {
                "group_id": {
                    "type": "string",
                    "example": "e933372f2147..."
                },
                "path": {
                    "type": "string",
                    "example": "/home/user/photo.png"
                }
            }
        },
...
        "routes.statusOK": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
...
    }
}